Extract review management into separate class
