+++ AutoML.py
@@ -1,0 +1,4 @@
+"""Compatibility module exposing AutoML application classes."""
+from mainappsrc.AutoML import AutoMLApp
+
+__all__ = ["AutoMLApp"]
